Add to directory service authorization, add settings file
