+++ pyproject.toml
@@ -13,6 +13,7 @@ uvicorn = {extras = ["standart"], version = "^0.30.6"}
 python-jose = {extras = ["cryptography"], version = "^3.3.0"}
 python-multipart = "^0.0.9"
 passlib = {extras = ["bcrypt"], version = "^1.7.4"}
+pydantic-settings = "^2.4.0"
 
 [tool.poetry.group.dev.dependencies]
 ruff = "^0.6.4"
