+++ src/main.py
@@ -3,10 +3,11 @@ from typing import Annotated
 from fastapi import Depends, FastAPI, HTTPException, status
 from sqlalchemy.orm import Session
 
+from src import auth
 from src.db.database import SessionLocal, database, engine  # noqa: F401
 from src.db.models.user import User  # noqa: F401
-from src import auth
-from src.auth import get_current_user
+from src.service.auth_service import get_current_user
+
 
 app = FastAPI()
 app.include_router(auth.router)
@@ -22,7 +23,7 @@ db_dependency = Annotated[Session, Depends(get_db)]
 user_dependency = Annotated[dict, Depends(get_current_user)]
 
 @app.get('/', status_code=status.HTTP_200_OK)
-async def user(user: user_dependency, db: db_dependency):
+async def user(user: user_dependency, db: db_dependency):  # noqa: ARG001
     if user is None:
         raise HTTPException(status_code=401, detail='Authentication Failed')
     return {'User': user}
