+++ src/settings.py
@@ -1,0 +1,7 @@
+from pydantic_settings import BaseSettings
+
+class Settings(BaseSettings):
+    SECRET_KEY: str = '12345'
+    ALGORITHM: str = 'HS256'
+
+settings = Settings()
